Fix lazy exposed modules
Lazy modules that were exposed by being external were not being handled
correctly. The require function would return the function that needed to
be called to return the exported object rather than the exported object
itself.

This commit adds lazy modules as getters, so that referencing one
calls the function and evaluates the module. It also adds a test case
based on the existing "expose" test case.

Fixes #40.